First round of splitting Lorre into functional modules
 * create custom processing operations modules for blocks, accounts,
   delegates, and one for the main app itself
 * extract the configuration read from lorre main program
 * separate Operations modules creation from the main body
 * align FeeOperations design to work on IO
